requirements_global.txt
numpy
pandas
pdfminer.six
opencv-python
lxml
xlrd
Pillow
flask
xlsxwriter
PyPDF2
requests
openpyxl
diff-match-patch
xlsxwriter
PyPDF2
pdfrw
comtypes
waitress
scikit-learn
flask-cors
flask-pymongo